getting rid of useActivityViewParams()
