+++ src/domain/shared/components/ActivityLog/views/ActivityChallengeCreatedView.tsx
@@ -5,8 +5,7 @@ import { useTranslation } from 'react-i18next';
 import { buildChallengeUrl } from '../../../../../common/utils/urlBuilders';
 import { NameableEntity } from '../../../types/NameableEntity';
 import { Caption } from '../../../../../core/ui/typography';
-import useActivityViewParams from './useActivityViewParams';
-import ActivityDescription from '../../ActivityDescription/ActivityDescription';
+import ActivityDescriptionByType from '../../ActivityDescription/ActivityDescriptionByType';
 
 export interface ActivityChallengeCreatedViewProps extends ActivityViewProps {
   challenge: NameableEntity;
@@ -31,24 +30,20 @@ export const ActivityChallengeCreatedView: FC<ActivityChallengeCreatedViewProps>
   });
   const url = buildChallengeUrl(journeyLocation.spaceNameId, challenge.nameID!);
 
-  const { i18nKey, values, components } = useActivityViewParams({
-    activityType: 'challenge-created',
-    author,
-    createdDate,
-    journeyTypeName,
-    journeyLocation,
-    parentDisplayName,
-  });
-
   return (
     <ActivityBaseView
       author={author}
       loading={loading}
       title={
-        <ActivityDescription
-          i18nKey={i18nKey}
-          values={values}
-          components={components}
+        <ActivityDescriptionByType
+          activityType="challenge-created"
+          {...{
+            author,
+            createdDate,
+            journeyTypeName,
+            journeyLocation,
+            parentDisplayName,
+          }}
           withLinkToParent={Boolean(journeyTypeName)}
         />
       }
